moved storage types to roles crate. renamed storage to engine.
