+++ node/components/executor/Cargo.toml
@@ -15,7 +15,7 @@ zksync_consensus_bft.workspace     = true
 zksync_consensus_crypto.workspace  = true
 zksync_consensus_network.workspace = true
 zksync_consensus_roles.workspace   = true
-zksync_consensus_storage.workspace = true
+zksync_consensus_engine.workspace = true
 zksync_consensus_utils.workspace   = true
 zksync_protobuf.workspace          = true
