+++ node/components/network/Cargo.toml
@@ -13,7 +13,7 @@ version.workspace    = true
 zksync_concurrency.workspace       = true
 zksync_consensus_crypto.workspace  = true
 zksync_consensus_roles.workspace   = true
-zksync_consensus_storage.workspace = true
+zksync_consensus_engine.workspace = true
 zksync_consensus_utils.workspace   = true
 zksync_protobuf.workspace          = true
