+++ node/libs/roles/src/validator/messages/v1/consensus.rs
@@ -1,4 +1,3 @@
-//! Messages related to the consensus protocol.
 use std::hash::Hash;
 
 use anyhow::Context as _;
