+++ node/libs/roles/src/validator/messages/v2/mod.rs
@@ -4,6 +4,7 @@ mod leader_proposal;
 mod replica_commit;
 mod replica_new_view;
 mod replica_timeout;
+mod state;
 /// Test-only utilities.
 mod testonly;
 #[cfg(test)]
@@ -15,3 +16,4 @@ pub use leader_proposal::*;
 pub use replica_commit::*;
 pub use replica_new_view::*;
 pub use replica_timeout::*;
+pub use state::*;
