--- node/libs/storage/build.rs
@@ -1,12 +1,0 @@
-//! Generates rust code from protobufs.
-fn main() {
-    zksync_protobuf_build::Config {
-        input_root: "src/proto".into(),
-        proto_root: "zksync/storage".into(),
-        dependencies: vec!["::zksync_consensus_roles::proto".parse().unwrap()],
-        protobuf_crate: "::zksync_protobuf".parse().unwrap(),
-        is_public: false,
-    }
-    .generate()
-    .expect("generate()");
-}
