+++ node/libs/storage/src/lib.rs
@@ -1,13 +1,13 @@
 //! Abstraction for persistent data storage.
 //! It provides schema-aware type-safe database access.
 mod block_store;
+mod interface;
 pub mod proto;
-mod replica_store;
 pub mod testonly;
 #[cfg(test)]
 mod tests;
 
 pub use crate::{
     block_store::{BlockStore, BlockStoreRunner, BlockStoreState, Last, PersistentBlockStore},
-    replica_store::{ChonkyV2State, Proposal, ReplicaState, ReplicaStore},
+    interface::{ChonkyV2State, Proposal, ReplicaState, ReplicaStore},
 };
