--- node/libs/storage/src/proto/mod.rs
@@ -1,2 +1,0 @@
-#![allow(warnings)]
-include!(concat!(env!("OUT_DIR"), "/src/proto/gen.rs"));
