--- node/libs/storage/src/types.rs
@@ -1,1 +1,0 @@
-
