+++ node/tools/Cargo.toml
@@ -18,7 +18,7 @@ zksync_consensus_crypto.workspace   = true
 zksync_consensus_executor.workspace = true
 zksync_consensus_network.workspace  = true
 zksync_consensus_roles.workspace    = true
-zksync_consensus_storage.workspace  = true
+zksync_consensus_engine.workspace  = true
 zksync_consensus_utils.workspace    = true
 zksync_protobuf.workspace           = true
